refactor : client - server component 분리
